simplify the code
