+++ GithubRepositoriesTaskTests/TestExtensions.swift
@@ -6,12 +6,25 @@
 //  Copyright © 2020 Mohamed Emad. All rights reserved.
 //
 
-import Foundation
+import XCTest
 
 extension URL {
 
     @inlinable static var testURL: URL {
         URL(string: .baseUrl)!
+    }
+    
+}
+
+
+extension XCTestCase {
+
+    var mockedData: Data? {
+        let testBundle = Bundle(for: type(of: self))
+        let filepath = testBundle.path(forResource: "LocalRepositories", ofType: "txt")!
+        let url = URL(fileURLWithPath: filepath)
+        let mockedData = try? Data(contentsOf: url)
+        return mockedData
     }
     
 }
